Split Signer interface and DefaultSigner class
